requirements.txt
Django==3.0.6
django-filter==2.2.0
django-import-export==2.1.0
django-tables2==2.3.1
numpy==1.18.3
pandas==1.0.3
requests==2.22.0
tablib==2.0.0
mysqlclient==1.4.6
